cambios 1

+++ src/main/java/com/example/Proyecto/Repositorio/RepositorioEmpleados.java
@@ -4,6 +4,6 @@ import com.example.Proyecto.Entidades.Empleados;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface RepositorioEmpleados extends JpaRepository<Empleados, Long>{
+public interface RepositorioEmpleados extends JpaRepository<Empleados, Integer>{
     
 }
